deps: remove langchain
Remove langchain and the associated pipeline to simplify our codebase
and reduce our exposure to problem coming from third party dependencies.
