+++ pyproject.toml
@@ -29,10 +29,6 @@ dependencies = [
   'ipython~=8.10.0',
   'jwcrypto~=1.5.6',
   'jinja2~=3.1.6',
-  'langchain~=0.3.10',
-  'langchain-core~=0.3.81',
-  'langchain-ollama~=0.3.5',
-  'langchain-text-splitters~=0.3.11',
   'launchdarkly-server-sdk~=8.3.0',
   'llama-stack-client==0.2.22',
   'protobuf~=5.29.5',
@@ -122,7 +118,6 @@ include = [
   "ansible_ai_connect/ai/api/aws/wca_secret_manager.py",
   "ansible_ai_connect/ai/api/model_pipelines/dummy/pipelines.py",
   "ansible_ai_connect/ai/api/model_pipelines/http/pipelines.py",
-  "ansible_ai_connect/ai/api/model_pipelines/langchain/pipelines.py",
   "ansible_ai_connect/ai/api/model_pipelines/llamacpp/pipelines.py",
   "ansible_ai_connect/ai/api/model_pipelines/ollama/pipelines.py",
   "ansible_ai_connect/ai/api/model_pipelines/wca/pipelines_base.py",
